requirements.txt
flask
torch
torchvision
opencv-python
pillow
numpy
matplotlib
scikit-learn
flask-cors
